refactured games code
