+++ bin/brain-games.js
@@ -1,4 +1,4 @@
 #!/usr/bin/env node
-import greetings from '../src/cli.js';
+import greetUser from '../src/cli.js';
 
-greetings();
+greetUser();
