+++ src/cli.js
@@ -1,23 +1,10 @@
 import readlinesync from 'readline-sync';
 
-/*
-  1. Здесь можно было бы поменять нейминг. Сейчас название функции не очень понятное, обычно лучше всего использовать глаголы. Например, greetUser. Но...
-  
-  2. ... эта функция делает две вещи одновременно: выводит приветствие в консоль и получает/возвращает имя юзера.
-  В небольшом коде вроде бы все нормально, в реальной разработке лучше так не делать, можно очень быстро запутаться)
-
-  3. Иногда создание целого файла для отдельной функции оправданно, но только если она очень большая (строк на 100), либо очень маленькая и переиспользуемая (такие называют утилитами).
-  Каноничный пример утилиты, который часто можно увидеть в проектах:
-  const capitalize = (string) => string[0].toUpperCase() + string.slice(1);
-  В данном случае, у нас ни то, ни другое.
-
-  По итогу, я бы предложил вообще не делать для этого отдельную функцию и отдельный файл, а унести эту часть прямо в engineGames.
-*/
-const greetings = () => {
+const greetUser = () => {
   console.log('Welcome to the Brain Games!');
   const name = readlinesync.question('May I have your name?');
   console.log(`Hello, ${name}!`);
   return name;
 };
 
-export default greetings;
+export default greetUser;
